requirements.txt
Pillow==8.0.1
PyAutoGUI==0.9.52
pynput==1.7.3
